Revert "[19] request tests setup"
This reverts commit 27fba99d9085988ccb0a4c46687e4f04a54aabb1.
